services/scraper/Auto Scraper/coder/agent.py
"""
Coder Agent - Generates Python web scraping code.
"""

from google.adk.agents import LlmAgent


INSTRUCTION = """
### ROLE
You are an expert Python developer specializing in web scraping.

### GOAL
Create a Python web scraper based on the Selector Map provided by the analyst.
The code must be production-ready, handle errors gracefully, and return structured data.

### REQUIREMENTS
1. **Function Signature:**
   - Create a function called `scrape(url: str) -> dict`
   - The function must accept a URL and return a dictionary with the scraped data

2. **Libraries:**
   - Use `requests` for fetching pages
   - Use `beautifulsoup4` (bs4) for parsing HTML
   - Include proper error handling

3. **Error Handling:**
   - Wrap requests in try-except blocks
   - Handle missing elements gracefully
   - Proper null checks:
     * Use `if elem:` before calling methods on elements
     * Never slice None (check `if text:` before `text[0:50]`)
     * Never iterate over None (check `if items:` before `for item in items`)
     * **CRITICAL**: Always add ALL fields to the output dict, even if they are None
     * **CRITICAL**: Always return an empty dict `{}` on error, NEVER return `None`
     * Add print outs across the code to help identify the source of any issue.

4. **Output Format:**
   - Return ONLY the Python code, no explanations
   - The code must be immediately executable
   - Include docstrings
   - Add debug print statements for each extracted field
   - Use proper null checks with the pattern: `elem.get_text(strip=True) if elem else None`
   - **DO NOT** include `if __name__ == '__main__':` block - the validator will add this automatically
"""


def create_coder_agent():
    """
    Creates and returns the coder agent.
    
    Returns:
        LlmAgent: Configured coder agent
    """
    return LlmAgent(
        name="coder",
        model="gemini-2.0-flash-exp",
        instruction=INSTRUCTION,
        tools=[]
    )
