services/scraper/Auto Scraper/requirements.txt
google-genai>=1.43.0
google-adk
requests
beautifulsoup4
lxml
